+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Changes:
 * New icons for public and private state (repos and gists)
 * Finetuned the icon. Thanks @benjaminrabe
 * Improved code highlighting
+* Improved repository view header
 
 Bugfixes: